requirements.txt
langchain==0.1.12
langchain-community==0.0.27
langchain-core==0.1.27
langchain-google-genai==0.0.11
python-dotenv==1.0.1
pypdf==4.1.0
streamlit==1.32.2
google-generativeai==0.3.2
faiss-cpu==1.7.4
sentence-transformers==2.5.1
torch==2.2.1
transformers==4.38.2